Added support for typescript
