+++ addons/docs/src/lib/type-system-handlers.test.ts
@@ -458,6 +458,7 @@ describe('prop-types handler', () => {
 
         const propDef = propTypesHandler(DEFAULT_PROP_NAME, docgenInfo);
 
+        expect(propDef.description).toBe('onClick description');
         expect(propDef.jsDocTags).toBeDefined();
         expect(propDef.jsDocTags.params).toBeDefined();
         expect(propDef.jsDocTags.params[0].name).toBe('event');
@@ -687,6 +688,7 @@ describe('prop-types handler', () => {
 
         const propDef = propTypesHandler(DEFAULT_PROP_NAME, docgenInfo);
 
+        expect(propDef.description).toBe('onClick description');
         expect(propDef.jsDocTags).toBeDefined();
         expect(propDef.jsDocTags.returns).toBeDefined();
         expect(propDef.jsDocTags.returns.description).toBe('An awesome string.');
@@ -701,6 +703,7 @@ describe('prop-types handler', () => {
 
         const propDef = propTypesHandler(DEFAULT_PROP_NAME, docgenInfo);
 
+        expect(propDef.description).toBe('onClick description');
         expect(propDef.jsDocTags).toBeDefined();
         expect(propDef.jsDocTags.returns).toBeDefined();
         expect(propDef.jsDocTags.returns.description).toBeNull();
@@ -745,6 +748,54 @@ describe('ts handler', () => {
     expect(propDef.description).toBe(docgenInfo.description);
     expect(propDef.required).toBe(docgenInfo.required);
     expect(propDef.defaultValue).toBe(docgenInfo.defaultValue.value);
+  });
+
+  it('should provide raw @param tags', () => {
+    const docgenInfo = createTypeScriptDocgenInfo({
+      tsType: PROP_TYPE_FUNC_TYPE,
+      description:
+        'onClick description\n@param {SyntheticEvent} event - Original event.\n@param {string} value',
+    });
+
+    const propDef = tsHandler(DEFAULT_PROP_NAME, docgenInfo);
+
+    expect(propDef.description).toBe('onClick description');
+    expect(propDef.jsDocTags).toBeDefined();
+    expect(propDef.jsDocTags.params).toBeDefined();
+    expect(propDef.jsDocTags.params[0].name).toBe('event');
+    expect(propDef.jsDocTags.params[0].description).toBe('Original event.');
+    expect(propDef.jsDocTags.params[1].name).toBe('value');
+    expect(propDef.jsDocTags.params[1].description).toBeNull();
+  });
+
+  it('should provide raw @returns tags when a description is defined', () => {
+    const docgenInfo = createTypeScriptDocgenInfo({
+      tsType: PROP_TYPE_FUNC_TYPE,
+      description:
+        'onClick description\n@param {SyntheticEvent} event - Original event.\n@returns {string} - An awesome string.',
+    });
+
+    const propDef = tsHandler(DEFAULT_PROP_NAME, docgenInfo);
+
+    expect(propDef.description).toBe('onClick description');
+    expect(propDef.jsDocTags).toBeDefined();
+    expect(propDef.jsDocTags.returns).toBeDefined();
+    expect(propDef.jsDocTags.returns.description).toBe('An awesome string.');
+  });
+
+  it('should provide raw @returns tags when there is no description', () => {
+    const docgenInfo = createTypeScriptDocgenInfo({
+      tsType: PROP_TYPE_FUNC_TYPE,
+      description:
+        'onClick description\n@param {SyntheticEvent} event - Original event.\n@returns {string}',
+    });
+
+    const propDef = tsHandler(DEFAULT_PROP_NAME, docgenInfo);
+
+    expect(propDef.description).toBe('onClick description');
+    expect(propDef.jsDocTags).toBeDefined();
+    expect(propDef.jsDocTags.returns).toBeDefined();
+    expect(propDef.jsDocTags.returns.description).toBeNull();
   });
 });
